[Server] Update VersionManager for 1.298 (#244)
* VersionManager: More consistent support for Unicode builds

* VersionManager: Disable unnecessary LS_MGAME_LOGIN handling

* VersionManager: Restore old news implementation

Also add SetString1() and SetString2() to make BYTE (1) vs short (2) length strings easier
and more efficient to set (rather than most of them calling strlen() twice).

* VersionManager: Restore updated server list behaviour

* VersionManager: Restore configuration defaults for ease-of-use

* VersionManager: Attach CVersionManagerDlg* through constructor

We could use AfxGetApp()->GetMainWnd() but this is overkill.
Just initialise it directly and we're fine.

The reason this doesn't work as-is is because AfxGetMainWnd()
fetches the window associated with the current thread.

Since packet requests are handled in the ReceiveWorkerThread which
is not associated with the window, it will fail.

* VersionManager: Use temporary manual auth logic for now

Ideally we should have this in its own account database, or at the very least, rename the procedure.
ACCOUNT_LOGIN as-is, is the game implementation. We do not want to use this for login as it is incompatible.

As we're currently using a singular database (and we expect people to be using our database), and we have no
means of syncing this currently, we'll temporarily hack this to fetch and handle basic auth logic
without a procedure.

* VersionManager: Restore premium lookup logic

Use LOAD_PREMIUM_SERVICE_USER as it's an account DB procedure.