Simplify and more explicitly build and parse news (#269)
* Simplify and more explicitly build and parse news

Officially, while the format is this clunky, news handling is not this explicit.
It just lazily finds and works around the #'s, ignoring the rest of it (e.g. the null-terminators)
and stripping the newlines (or more specifically, both \r and \n).

This "works" officially because it doesn't render the null-terminators it's included, but this is a silly
detail that only works by happenstance.
We'll just keep expecting it to be in the given format no matter what.

I also just simplified the sequences to explicitly state the start, and end of the message blocks themselves.
Not the news blocks in their entirety, since titles are handled outside.

The notion of LINE_ENDING was just my old assumption based on the pattern of the format,
but given the official parsing, it has no real basis in reality.
This was merged into the other sequences appropriately.

Additionally, if there's no news, it should just skip straight to the server list.

Resolves #249