Add ThirdParty.sln and manually build zlib (#284)
* Remove precompiled zlib libs

* Add zlib 1.2.7 as a submodule

This is the version we are currently using.
We can update it later, but for now the focus is to just shift things.

* Add ThirdParty.sln for "one time" building

Also update the Launcher & VersionManager to reference it correctly,
as well as the various workflows to lazily fetch the submodule and always
build ThirdParty.sln first (even if it may not necessarily need it - it will be fast).

* Rename update_submodules.cmd -> update_dependencies.cmd, add build logic

In addition to initializing and updating the submodules, this will now also find msbuild.exe and build
ThirdParty.sln for both Debug & Release builds.

This information has been added to README.md

* Specifically reference the potential linker error as well