Remove N3Base client dependencies (#297)
* CN3Cloak: Remove dependency to WarFare project

Technically this isn't even called, but when it is it would just map
the game's move state to the cloak's move state, as well as passing it
the character's current yaw.

* Move CN3UIWndBase::m_sRecoveryJobInfo.m_bWaitFromServer to CN3UIBase::s_bWaitFromServer

This detaches it from the CN3UIWndBase dependency and all that entails.

* CN3UIArea: Move inventory-specific logic back to CUIInventory

Also disable UIMSG_AREA_DOWN_FIRST -- it sends UIMSG_BUTTON_CLICK for everything
except the inventory normally, so we should just use that instead since it's easier
to remove the 1 now unused case vs tracking down all the UIMSG_BUTTON_CLICK cases.

* CN3ShapeMgr: Restore load behaviour & don't depend on client classes