Chunk MP3 decoder writes instead of predicting and writing in one go (#375)
The prediction was off, but it's just wasted memory regardless. The original intent was to reuse this logic for loading the mmio buffer from memory, but after looking more at the interfaces, I never bothered to go down that route; it wasn't worth it.

So we'll conserve memory and avoid unintentional overruns here from bad prediction code by just writing back out in chunks.

Closes #372