Rename server table and related maps/containers for consistency
It's a little verbose, but we no longer use 'Array' to denote maps (we use 'Map' instead), and only continue to use 'Array' for vectors since it's the same basic concept),
but we always explicitly indicate whether it's a table container or not via 'Table'.
This is because of cases like the NPC map; there's an NPC table (i.e. the NPC table map) and
then there's the regular NPC (instance) map.