Require a minimum Visual Studio version for msbuild (VS2022)
It needs to support v143 (2022).
If it's between a preview and mainline build, they NEED to be 2022+.

e.g. Mainline build: 2019 build tools
Preview build: 2022

We prefer the mainline build, so previously it would select the 2019 build tools.
Now it will still ignore the mainline build if it's not VS2022 or newer (v17.0+).