ByteBuffer: Force explicit reads via read<T>()
Stream operators are nice, but particularly for reads, they mask problems here.
If a type externally changes, there's no way to enforce or make it immediately clear what type we're reading.

We should just be explicit about it so that the protocol is immediately obvious and consistent.

Writing is another story. I leave the >> operators for now; it may be worth replacing them, but this issue is
mitigated simply by enforcing the cast when using the operator, e.g.:
pkt << uint16_t(field);

This makes it immediately obvious what the protocol is and avoids it breaking when 'field' changes externally,
though code-wise it obviously isn't enforced. It also ends up not obviously pairing with the reads, but I think
it's still a step in the right direction for enforcing consistent protocol behaviour.