Add unofficial internal cooldown to establishing game connections (#431)
As-is, the client can spam successful game server connections as there's a window between it connecting and sending the WIZ_VERSION_CHECK packet, and it getting a response. In this window, the player can initiate another request, causing it to establish another connection. If the WIZ_VERSION_CHECK packet from the first connection is received, it ends up associating the wrong encryption key -- because that's no longer the session it's using it for.

Notably this is, for the most part, official behaviour. It is very possible to continue to spam the game server with the official client, but they force a 1s delay between connection requests which limits the frequency you can actually spam. The bug is still very much an issue there though.

The official Sleep() approach isn't ideal; it doesn't really solve a lot. We should simply be unable to establish a new game server connection while we're waiting for the WIZ_VERSION_CHECK response. This is unofficial behaviour of course, but it doesn't really affect compatibility between clients (only its connection behaviour for testing). As it is, this causes (more) unintentional disconnects at login, which I'd argue makes it harder to even get to testing 99% of other behaviour, so it should still be improved.

The official Sleep() approach doesn't really prevent the inconsistency of the unintentional "failures". All it does is it stops the game server from being excessively flooded with new connections, since it throttles it to 1 a second.

Instead, we enforce a 5 second internal cooldown on ConnectToGameServer() calls. 5 seconds is plenty of time to receive a packet, even with lag. This is reset if we detect a disconnect, and the timer otherwise ticked down in GameProcLogIn_1098::Tick() / CGameProcLogIn_1298::Tick().

This is further reset when we return back to the login scene (on Init()), so returning and selecting a new server doesn't force the player to unnecessarily wait.

This prevents us from accidentally or otherwise attempting to connect multiple times, causing mixed WIZ_VERSION_CHECK request handling and thus incorrect encryption states/keys from being used.

Resolves #430